+++ util/doc_gen.py
@@ -14,6 +14,7 @@ if __name__ == "__main__":
              "transform.py",
              "wheel.py",
              "test_controller.py",
-             "drive.py"]
+             "drive.py",
+             "turn_constants.py"]
     md = PyMdDoc(input_directory="magnebot", files=files, metadata_path="doc_metadata.json")
     md.get_docs(output_directory="doc/api")